backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
celery>=5.3
redis>=5.0
requests>=2.31
beautifulsoup4>=4.12
openai>=1.0
openpyxl>=3.1
python-dotenv>=1.0
gunicorn>=21.2
drf-spectacular>=0.27
google-generativeai>=0.8
